requirements.txt
accelerate==0.33.0
albumentations==1.4.14
anthropic==0.34.1
bayesian-optimization==1.5.1
bayespy==0.5.1
biopython==1.84
catboost==1.2.5
cython==3.0.11
dacite==1.8.1
datasets==2.1.0
efficientnet-pytorch==0.7.1
evaluate==0.4.2
fastai==2.7.17
faust-cchardet==2.1.19
gym==0.26.2
h5py==3.11.0
hmmlearn==0.3.2
hyperopt==0.2.7
imagecodecs==2024.6.1
imbalanced-learn==0.12.3
imgaug==0.4.0
implicit==0.7.2
ipykernel==6.29.5
ipython==8.27.0
jsonlines==4.0.0
kaggle==1.6.17
keras==3.5.0
kornia==0.7.3
langchain-anthropic==0.1.23
langchain==0.2.15
levenshtein==0.25.1
librosa==0.10.2.post1
lightgbm==4.5.0
markovify==0.9.4
matplotlib==3.9.2
nbformat==5.10.4
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
ogb==1.3.6
openai==1.103.0
opencv-python==4.10.0.84
optuna==4.0.0
pandas==2.2.2
pdf2image==1.17.0
peft==0.12.0
pillow==10.4.0
plotly==5.24.0
py7zr==0.22.0
pyarrow==17.0.0
pydicom==2.4.4
pyocr==0.8.5
pyparsing==3.1.4
pypdf==4.3.1
pytest==8.3.2
pytorch-lightning==2.4.0
pytz==2024.1
pyyaml==6.0.2
ranger21==0.1.0
resampy==0.4.3
rouge-score==0.1.2
sacrebleu==2.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scikit-optimize==0.10.2
scikit-surprise==1.1.4
scipy==1.14.1
seaborn==0.13.2
segmentation-models-pytorch==0.3.4
sentence-transformers==3.0.1
sentencepiece==0.2.0
sklearn-pandas==2.2.0
spacy==3.7.6
sympy==1.13.2
tensorflow-hub==0.16.1
tensorflow==2.17.0
tensorpack==0.11
tf-keras==2.17.0
tiktoken==0.7.0
timm==0.9.7
torch-geometric==2.3.1
torch==2.2.0
torchaudio==2.2.0
torchinfo==1.8.0
torchmetrics==1.4.1
torchtext==0.17.0
torchvision==0.17.0
tqdm==4.66.5
transformers==4.44.2
xgboost==2.1.1
xlrd==2.0.1
python-dotenv==1.1.0
litellm==1.76.1
nbclient==0.10.2
diskcache==5.6.3
appdirs==1.4.4
optuna-integration==4.5.0
deepchem==2.8.0
rdkit==2025.9.3

# conda create -n ds-lumina python=3.11 -y
